auth_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
py_eureka_client